Fix gomod auth implementation
The auth header needs to be:
```
Authorization: basic <token-hash>
```

Where `<token-hash>` is the Base64 encoding of the string
`"x-access-token:<token>"` and `<token>` is the token as provided by the
authenticator (usually github). Prior commits missed the
`x-access-token` part. It still worked, but was not strictly correct.

Some other minor implementation details were changed, such as when shell
quotes are added.

This commit also adds some tests, but they will not succeed. Certain
files are embedded as strings, and those files were not checked in for
security reasons. The test implementation is also not correct, so it
needs to be changed anyway.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>