Modify earlier test
The thought process was to try to write a test that checks the generated
gitConfig. But that's not really an integration test, it's more of a
unit test. The proper thing to do would be to check the llb returned
from the StateOption -- maybe even solve it and look at the result.

As such, this commit will be reverted.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>